unified lookup results, started qrm.py, some linting
