+++ commands/morse.py
@@ -5,8 +5,10 @@ from core import bot, to_morse
 '''
 Translates from text to periods and dashes
 '''
+
+
 @bot.command()
-async def morse(ctx, *, text:str):
+async def morse(ctx, *, text: str):
     morsemess = ''
     for i in text:
         if i in to_morse:
