+++ onlinelookup/qrm.py
@@ -1,0 +1,17 @@
+"""
+QRM Callsign Lookup
+Author: Ben Johnson, AB3NJ
+
+Uses qrm.website's API to retreive information on callsigns
+"""
+
+
+import json
+import os.path
+from urllib import request, error
+from . import olerror
+
+
+class QRM:
+    def lookup(self, callsign, params):
+        pass
